Move layer styles and sql into subfolders
