+++ flex-config/run-all.lua
@@ -1,2 +1,2 @@
-require "tags"
+require "style.tags"
 require "run-no-tags"
