+++ flex-config/run-all.sql
@@ -1,2 +1,2 @@
-\i tags.sql
+\i sql/tags.sql
 \i run-no-tags.sql
