+++ flex-config/run-no-tags.lua
@@ -1,13 +1,13 @@
-require "pgosm-meta"
-require "amenity"
-require "building"
-require "indoor"
-require "infrastructure"
-require "landuse"
-require "leisure"
-require "natural"
-require "place"
-require "road"
-require "shop"
-require "traffic"
-require "water"
+require "style.pgosm-meta"
+require "style.amenity"
+require "style.building"
+require "style.indoor"
+require "style.infrastructure"
+require "style.landuse"
+require "style.leisure"
+require "style.natural"
+require "style.place"
+require "style.road"
+require "style.shop"
+require "style.traffic"
+require "style.water"
