+++ flex-config/run-no-tags.sql
@@ -1,13 +1,13 @@
-\i pgosm-meta.sql
-\i amenity.sql
-\i building.sql
-\i indoor.sql
-\i infrastructure.sql
-\i landuse.sql
-\i leisure.sql
-\i natural.sql
-\i traffic.sql
-\i place.sql
-\i road.sql
-\i shop.sql
-\i water.sql
+\i sql/pgosm-meta.sql
+\i sql/amenity.sql
+\i sql/building.sql
+\i sql/indoor.sql
+\i sql/infrastructure.sql
+\i sql/landuse.sql
+\i sql/leisure.sql
+\i sql/natural.sql
+\i sql/traffic.sql
+\i sql/place.sql
+\i sql/road.sql
+\i sql/shop.sql
+\i sql/water.sql
