+++ flex-config/run-unitable.lua
@@ -1,0 +1,2 @@
+require "style.pgosm-meta"
+require "style.unitable"
