+++ flex-config/run-unitable.sql
@@ -1,0 +1,2 @@
+\i sql/pgosm-meta.sql
+\i sql/unitable.sql
