+++ flex-config/style/leisure.lua
